+++ src/main/java/com/verdantartifice/primalmagick/common/config/Config.java
@@ -19,7 +19,6 @@ import net.minecraftforge.fml.loading.FMLPaths;
  */
 @Mod.EventBusSubscriber
 public class Config {
-    protected static final String CATEGORY_WORLDGEN = "worldgen";
     protected static final String CATEGORY_MISC = "misc";
     
     protected static ForgeConfigSpec COMMON_CONFIG_SPEC;
@@ -28,9 +27,6 @@ public class Config {
     
     public static ForgeConfigSpec.BooleanValue SHOW_AFFINITIES;
     public static ForgeConfigSpec.BooleanValue SHOW_UNSCANNED_AFFINITIES;
-    public static ForgeConfigSpec.BooleanValue GENERATE_MARBLE;
-    public static ForgeConfigSpec.BooleanValue GENERATE_ROCK_SALT;
-    public static ForgeConfigSpec.BooleanValue GENERATE_QUARTZ;
     public static ForgeConfigSpec.EnumValue<TheorycraftSpeed> THEORYCRAFT_SPEED;
     
     static {
@@ -41,12 +37,6 @@ public class Config {
     protected static void buildCommonConfigSpec() {
         // Define the common config file spec
         ForgeConfigSpec.Builder builder = new ForgeConfigSpec.Builder();
-        
-        builder.comment("Worldgen settings").push(CATEGORY_WORLDGEN);
-        GENERATE_MARBLE = builder.comment("Whether to generate raw marble in the Overworld.", "WARNING: Disabling this will cripple players' ability to progress", "without another source of this mod's marble.").define("generateMarble", true);
-        GENERATE_ROCK_SALT = builder.comment("Whether to generate rock salt ore in the Overworld.", "WARNING: Disabling this will cripple players' ability to progress", "without another source of this mod's salt.").define("generateRockSalt", true);
-        GENERATE_QUARTZ = builder.comment("Whether to generate quartz ore in the Overworld.", "Disabling this will slow players' ability to progress.").define("generateQuartz", true);
-        builder.pop();
         
         builder.comment("Misc settings").push(CATEGORY_MISC);
         SHOW_UNSCANNED_AFFINITIES = builder.comment("Show affinities of blocks and items even without scanning them").define("showUnscannedAffinities", false);