+++ changelog.md
@@ -1,6 +1,8 @@
 v3.0.0-beta
 -----------
 * Updated to Minecraft 1.19
+* Removed vestigial worldgen config options
+    * NOTE: Generation of marble, salt, and quartz is now controlled via datapacks
 
 v2.1.2
 ------